+++ R/01-includeDataToolsCSS.R
@@ -1,0 +1,6 @@
+#' Include CSS from DataTools package
+#'
+#' @return NULL
+includeShinyToolsCSS <- function() {
+  tags$link(href = "app_files/custom.css", rel = "stylesheet")
+}