+++ R/01-onLoad.R
@@ -1,0 +1,9 @@
+#' Add path to css
+#'
+#' @param ... further arguments
+#'
+#' @return NULL
+.onLoad <- function(...) {
+  addResourcePath(prefix = "app_files",
+                  directoryPath = system.file("app", "www", package = "DataTools"))
+}